Add best effort support for locating WinSDK tools
For platforms that don't have VS2017 or newer installed the vswhere based solution for locating the WinSDK dir doesn't work.  This change adds support for a best effort to look in known default install directories for the WinSDK tools.   The prefered method is still to use VsWhere and the vsvars files but this can help on older platforms.