Add best effort support for locating WinSDK tools (#3)
* Add best effort support for locating WinSDK tools
* Add support for case where FindWithVsWhere returns None.
* Add Unittests to validate behavior

For platforms that don't have VS2017 or newer installed the vswhere.exe based solution for locating the WinSDK directory doesn't work.  This change adds support for a best effort to look in known default install directories for the WinSDK tools.   The preferred method is still to use vswhere and the vcvars.bat files but this can help on older platforms.