base_parser: Fix DSC !ifdef and !ifndef parsing
The syntax of an !ifdef and !ifndef statement does not
use the value of a define that uses $().  Instead, only
the name of the define is used.

Update the test for !ifdef and !ifndef to check if the
define name in token[1] is in LocalVar or not and push
the correct conditional value of True or False.

This change was tested locally with the edk2/NetworkPkg.
With this change, the defines "Defines" section of
NetworkPkg.ci.yaml can be emptied, and the default values
for all defines are correctly parsed.

Signed-off-by: Kinney <michael.d.kinney@intel.com>