edk2toollib/base_parser: Fix IN operator and quoted arguments
Update the Base Parser to support a DSC file conditional
statement such as the following:

!if $(CRYPTO_SERVICES) IN "ALL NONE MIN_PEI MIN_DXE_MIN_SMM"

One of the arguments is a quoted string that needs to be
parsed as a single argument.

The IN operator tests for membership.  Update parser to support
upper a lower case forms of the IN operator and use the
python set() type to test for membership if the IN operator
is being evaluated.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>