Comments switched to docstring, and inline documentation improved
FirmwarePolicy().Serialize() now takes a bytearray, not a stream, as a
parameter, making it consistent with all other Serialize() methods in
this family.  A SerializeToStream() is provided for convenience that
writes the serialized value to a provided stream.

Added a testcase for optimized policies where multiple rules reference
the same SubKeyOffset and/or ValueNameOffset (because the contents are
identical).
Added testcases for some boundary and error cases
Code coverage of firmware_policy.py is 91%