.vscode/settings.json: Add workspace settings for unit tests
Adds a VS Code workspace settings file to define the
"python.testing.unittestArgs" for running Python unit tests.

This allows unit tests to automatically be recognized and shown
in VS Code Test Explorer.

Without this change, the tests are not shown in Test Explorer
and an error message is output to Output -> Python window due to
the test root directory not being valid.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>