edk2toollib/uefi/edk2/path_utilities.py: WARNING instead of exception for nested package
Update the check for nested packages to generate a WARNING instead of
raising an exception.

This check is valuable and all nested packages should be removed.
By downgrading this check to a WARNING, the downstream consumers
of edk2 can be notified of the issue and be given time to resolve
without breaking stuart builds.

Disable the unit test for this expected exception until all
nested package issues can be resolved.

Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>