Edk2Path: Add an env variable to allow nested packages
Adds an environment variable called `STUART_IGNORE_EDK_NESTED_PACKAGES`
that a platform can set to `"true"` to allow nested packages.

Nested packages are a violation of EDK II package rules and should
be fixed. This is provided for backward compatibility and temporary
use until the variable is no longer needed.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>