Add support for ignoring known-bad packages (#227)
Adds an environment variable to allow the nested package check to be demoted to an INFO-level log for packages that are known to be bad and cannot be immediately remediated. This avoids a large number of warning messages that would otherwise make it more likely for developers to ignore warnings entirely.