utility_functions: MacOS bugfix
This commit fixes two bugs that cause unit tests to fail on MacOS. The
first is that tempfile uses a symlink when generating the temporary
directory, so the generated tempfile path is now resolved to account
for this symlink. The second is that MacOS subprocess.Popen() expects
the command to be passed as a list of argument / flags rather then a
string.