base_parser.py: Allow Edk2PathObj to be passed (#286)
base_parser.py: BaseParser was identified as a class that instantiates an Edk2Path object but could also be instantiated itself many times throughout the build process (once per INF). As a part of https://github.com/tianocore/edk2-pytool-extensions/issues/446, we are working to minimize the amount of Edk2PathObj instantiations due to the slow verification of workspace paths that occur.

This change adds a new method, `SetEdk2Path()` which is a drop in replacement for `SetBaseAbsPath()` and `SetPackagePaths()`.

Edk2Path's constructure requires both the workspace path (BaseAbsPath) and the Packages path list. Due to this, we can instantiate the Edk2Path object once, and pass it to each `BaseParser()` instance, changing the amount of calls from 2*num_of_parsers to 1.

Additionally, `SetBaseAbsPath()` and `SetPackagePaths()` have been deprecated.