parsers: improve existing parsers
base_parser.py: Add ability to handle elseif conditionals

dsc_parser.py: Generates a library dictionary which uses the dsc section
to scope libraries to a specific architecture and or module type.
dsc_parser.py: Adds a new attribute, Components, which stores Components
described in the dsc.
dsc_parser.py: Adds the ability to evaluate conditionals when the
the condition involves a PCD.

fdf_parser.py: When in a section, if a DEFINE statement is present, that
define is added to the LocalVars, which is used for conditional
evaluations.

inf_parser.py: Generates a library dictionary which uses the inf section
to scope libraries to a specific architecture.