locate_tools: set comparison invalid due to mismatched case of keys (#313)
Within `QueryVcVariables()`, A set comparison is made between the requested variables and the returned variables to determine (and raise an error printing) the missing keys. There is a difference in case for the two sets, one being all uppsercase values while the other is not. This change updates both sets to contain all uppercase values.