database: add pre-made queries
Adds the following pre-made queries that can be used without the need
of making custom queries:

ComponentQuery: A query that returns information about a component that
would be build by the parsed dsc based off the ENV values set at the
time of parsing.

LibraryQuery: A query that returns every instance of a particular
LIBRARY_CLASS found in the parsed workspace

LicenseQuery: A query that returns all files in the parsed workspace
that do not contain a valid SPDX license identifier.

UnusedComponentQuery: A query that returns all components and libraries
that would be built per a dsc, but not used by the FDF based off the
ENV values set at the time of parsing.