ansi_handler: support all string representable objects in LogRecord message (#379)
Previously, the ansi_handler only accepted strings in the logging record message due to a bug in how the message was retrieved from the record. This change fixes this, allowing any object that str() works on to be used as a message in the logging record.

This fix works by converting the message to a str (similar to LogRecord.getMessage()) then appending the color code. We cannot call .getMessage() directly as that will insert any args into the string, which causes a break later when .getMessage() is called again by the logging library (An exception is raised due to attempting to insert an argument into a string without a corresponding % placeholder).