base_parser: support relative path include files
This commit adds support for including multiple relative paths from
the same directory by storing a stack of all included files that have
not finished being parsed, and the number of lines that still need to
be parsed in each file.

Previously, only includes of nested relative paths were supported as
the currently parsed file was not reverted to the previous file when the
parsing of an included file was completed. This led to further includes
needing the be relative to the latest included file.