base_parser: condition eval bugfix
Fixes the bug where the `ComputeResult` attempts to evaluate the result
using the original string value, rather than the converted int value
when handling greater than or less than conditions.

It should have been using the ivalue, which is the conversion of the
parsing string into an actual integer (i.e. "1" to 1 or "0x0" to 0).