edk2toollib/uefi/edk2/parsers: enhancements (#419)
Improves the following parsers in the following ways:

`base_parser.py`: Properly uses ivalue (int value) to check if the string representation of a value was correctly converted to an integer ('1' to 1, '0x0' to 0, etc) when evaluating conditionals

`dsc_parser.py`: register pcds when performing the initial parse responsible for finding define statements. This is necessary to evaluate conditionals that use pcds.