dsc_parser.py: use SUPPORTED_ARCHITECTURES
to properly scope components to a specific architecture when they fell
under a common section, the parser relied on TARGET_ARCH being provided
as a InputVar via SetInputVars().  This came with two problems: (1) It
was not garunteed that TARGET_ARCH would be provided, which would result
in the component registered for any scope and (2) it limited the results
to particular architecture rather than the actual supported
architecture per the DSC.

This commit changes the parser to use the SUPPORTED_ARCHITECTURES define
as set in the DSC itself, which is a required DEFINE per the DSC
specification.