path_utilities: Remove Nested package exceptions
Removes Nested package exceptions from path_utilities.py in favor of a
warning.  This is being done as the exception has served its purpose and
those willing to update their nested packages have had a year to do so.
Any nested packages will now only be a warning logged at logging.DEBUG
and issues with converting from an absolute path to an edk2 package path
relative path will be accepted.

A new CI Check has been added so that packages managed by that specific
repository will still be checked by nested packages, but nested packages
outside of the control of that specific repository will not be checked
and thus not result in an exception being raised.