utility_functions: Add LongPath support to RemoveTree
On Windows systems with long paths disabled, the RemoveTree function
fails to remove files that go over the character limit as they cannot
be found. This change, on windows, prefixes the path with the '\\?\'
which allows for the detection of long paths, and allows for the removal
of the file.