table parsers: Use closest package path relative path (#461)
INF paths in the DSC and FDF can be defined as an absolute path, relative to the workspace, or relative to any package path. To ensure consistency with paths, this change converts any provided path to be relative to the closest package path. This simplifies queries that typically use the `path` column for filtering, or `join` commands.

That is to say, if there are two package paths: `Common/`, `Common/Special/`, a path can be relative to either of these packages paths. So both `Special/Files/MyFile.c` and `Files/MyFile.c` are acceptable. However when inserting into the database, we now ensure that the closest package path relative path is used. In the above example, that means `Files/MyFile.c`.