locate_tools: Search all VS installations for vcvarsall.bat
Adds a new function, FindAllWithVsWhere which functions the same as
FindWithVsWhere, but returns a list of all paths found, rather than the
first path only.

Updates QueryVcVariables to use FindAllWithVsWhere, and loops through
each path found, attempting to locate the vcvarsall.bat file. It will
use the first one found; if none are found, it will print an error and
raise an exception.