Convert Edk2DB to use an ORM instead of sqlite3 directly (#485)
Edk2DB has a high barrier to entry as users must not only know how to write and work with sqlite3 databases, but also spend the time to understand the schema of the EDK2DB database, which can change depending on user's needs. These changes occur because users can choose to register only a subset of provided parsers, and create their table parsers, which can result in a drastically different database implementation. This change switches database functionality to be managed through an ORM (specifically [https://www.sqlalchemy.org/](https://www.sqlalchemy.org/)) to abstract away the need to know the inner workings of the database.

This change should greatly reduce the barrier to entry for users that want to utilize this functionality. They only need to learn the simple interface for writing ORM queries (Through the [ORM Querying Guide](https://docs.sqlalchemy.org/en/20/orm/queryguide/index.html)), then they will be able to work with the database tables as if they were traditional python objects.

This change does the following:

1. Switches Edk2DB to use sqlalchemy instead of sqlite3 (sqlalchemy uses sqlite3).
2. Removes the Edk2DB context manager as it is no longer necessary
3. Add Mapped Table objects to `edk2toollib.database` which represent tables in the database
    - Environment
    - Value
    - Inf 
    - InstancedInf
    - Source
    - Library
    - Package
    - Fv
4. Updates all table generators to use sqlalchemy functionality to insert into the database
5. Updates documentation