buildreport_parser: Support component INF absolute paths (#520)
A path from an ext dep used for an INF will be absolute. This causes
the build report parser to fail to match that path to a component
since the check is performed against a relative path.

This change checks if a path is absolute before the comparison. If
so, the path is converted to a package path adjusted path for the
comparison.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>