Update OS version to 17134 to limit DDInstall sections (#541)
Limit capsule install inf changes to OS versions 17134 or higher for a newer HDC requirement.
Original PR - #431

"According to Zac Lockard for the InfVerif team, the new WHCP requirements went into force with the newest 26080 build which is why we're seeing this now. For reference, they link https://learn.microsoft.com/en-us/windows-hardware/drivers/devtest/inf-validation-errors-and-warnings#1199 and quote the page that says "You must fix all errors in order to pass driver tests in Hardware Dev Center."

It is the Device.DevFund.INF.Syntax requirement

This change was also pushed to all older versions of the OS due to this being a problem often in the ecosystem".