Documentation: Fix all broken links (#609)
[MkDocs v1.5.0](https://github.com/mkdocs/mkdocs/releases/1.5.0) Updated how links were evaluated (See "Expanded Validaton of links" section). Before 1.5.0, most links in our documentation were left "untouched" as they either started or ended with a `/`.

> ... links that started with `/` ("absolute") and links that _ended_ with `/` were left as is ...

However these links are now properly evaluated, which broke most links in edk2-pytool-extensions documentation. With this new evaluation, All of our links that started absolute (`/`) were now linking relative to `tianocore.org` instead of `tianocore.org/edk2-pytool-library`.

This pull request updates the documentation to utilize the newly introduced `absolute_links` validation setting to be `relative_to_docs` which updates absolute paths to be relative to `tianocore.org/edk2-pytool-extensions` and updates any path references to point to the markdown file, which is necessary for the `relative_to_docs` setting to work as expected.