Python support: drop 3.10, add 3.13 (#652)
## Breaking Change

This commit removes support for python 3.10. Attempting to install v0.22.0 of edk2-pytool-library using a python version less than 3.11 will result in an error from pip along the lines of `ERROR: Could not find a version that satisfies the requirement ...`

## Integration Instructions

If your platform cannot upgrade to 3.11 or above, continue using a version of edk2-python-library less than v0.22.0. Otherwise upgrade your version of platform to 3.11 or greater.