fdf_parser.py: support `SECTION COMPRESS` sections (#654)
Previously, the FDF parser would fail to parse an FDF file with statements containing a SECTION COMPRESS with additional with nested sub sections. 

This PR updates the parser to be able to get past the parsing of the this, by checking there is an equal in the statement line, and pass it.