base_parser.py: support conditions checking for empty string (#653)
base_parser.py did not support conditional directives that were comparing to empty strings (e.g. `!if $(MY_VAL) == ""`, `!if $(MY_VAL) != "". The logic in `_TokenizeConditional`, was setting `mode` to QUOTE_MODE twice because the token length was zero for an empty quote (`len(token) == 0`).

This PR Updates the logic to not enter QUOTE_MODE when its already in QUOTE_MODE, and to fall over to end of QUOTE_MODE.