Fix UEFI Variable Reading on Linux (#655)
The UEFI variable fetching on Linux was broken, it was reading
the attributes as part of the data, not correctly skipping over
the attributes and fetching the data. For Linux, the value read
is a UINT32 Attributes + Variable Length Data.

This change has been tested to fix that but reading (and
ignoring but logging an error if the attributes aren't 7, as
ConfigEditor expects) the attributes and then fetching the data
past that.