Update operator precedence for post fix evaluation
Equals '==' operator should have a higher precedence than OR and AND.
This fixes a bug where a conditional like
"!if RACECAR == RACECAR ||  RACECAR == STREETCAR" would evaluate the last two
operators in the wrong order, returning False when it should be True.
Removing extra parenthesis at the end of the tokens as it is modifying the
input object and is only needed because an opening one is added to the stack.