base_parser.py: Update precedence of `==` and `!=` (#684)
To align with industry standard order of operations, and with EDKII parsers, equals (`==`) and not equals (`!=`) should have a higher precedence than logical OR (`||`) and logical AND (`&&`) operators.

Prior to this commit, they shared the same precedence level which would lead to evaluations similar to `!if RACECAR == RACECAR || RACECAR == STREETCAR` being completed in the wrong order as compared to EDKII parsers. Additionally, this removes the parentheses that were inserted at the beginning and end of tokens as they appear to serve no further purpose.