base_parser: bugfix ifdef ifndef evaluation (#698)
Previously, ifdef / ifndef evaluation was done by replacing the macro with a "0" if it was not defined then when evaluating the conditional, it was just verified if the value was "0" or not. This left open a scenario where a macro is defined but defined as an empty string. In this scenario, the parser would successfully replace the macro value, but in replacing it with an empty string, if left bare conditionals (e.g. the line was only !ifdef or !ifndef, which would cause an exception when attempting to evaluate that conditional.

This commit updates the logic to only check if the macro is defined or not, replacing the value with a "0" if missing, or a "1" if existing. Now, regardless of the value of the macro, the conditional will always be valid.