project: Add a minimum supported version
Adds the concept of a minimum supported version (MSV) to the project.

The project will continue to always maintain and support N-2 versions of
Python, however the MSV will be the oldest version of python that can
run the project (i.e. the project does not use any features that are
newer than the MSV). This value is subject to change at any time due to
the needs of the project.