locate_tools: QueryVcVariables: Allow exact vc (#710)
Allows a caller to specify the exact compiler version to use for
querying variables. This is useful because a system may have a single
installation of (example) VS2022, but may have multiple versions of the
VS2022 compiler installed. Prior to this change, QueryVcVariables would
return variables arbitrarily based off of one of the installed
compilers.

This change allows the caller to specify the exact version of the
compiler to use for querying variables, so that the returned variables
are consistent.
ate