base_parser: Give input variables precedence to local variables
Right now, if a variable is given as an input variable (e.g. a cmd
param), and it has a local variable set to `FALSE`, the input
input variable is not considered.

At a higher level, this is causing the DSC parser to not consider
an input variable for a build value that is set to `TRUE` on the
command line, when it is defaulted to `FALSE` in the `[DEFINES]`
section of the DSC.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>