edk2_db: Fix mis-aligned zip for library dependencies (#744)
This commit fixes a bug that results in a zip of two different sized
arrays that must be the same size. This mis-alignment results in
recursive library class dependencies being associated with the incorrect
parent library or component.

This commit fixes the issue by creating two new lists and only ever
appending to both, where as previously a premature continue in a loop
had been shown to mis-align the lists.