dsc_parser: Some PCD value not working in !if directive
__ProcessDefines is getting called recursively for every !include
directive. But PcdValueDict is reset at the end of every instance of
__ProcessDefines, which means condition directives can only refer to PCDs
assigned in the same .dsc file which is an incorrect behavior.

Moving it out to the ParseFile function, after the very first call of
__ProcessDefines so we can refer to PCDs assigned previously in other
dsc file and all of them are cleared after processing Defines.

Also fix another issue that during process Defines, only the first PCD
in the condition directive will be parsed.

Signed-off-by: PaddyDeng <paddydeng@ami.com>